requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
Deprecated==1.2.14
distro==1.9.0
frozenlist==1.4.1
idna==3.8
magic-filter==1.0.12
multidict==6.0.5
netaddr==1.3.0
outline-vpn-api==6.3.0
pydantic==2.8.2
pydantic_core==2.20.1
python-decouple==3.8
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
wrapt==1.16.0
yarl==1.9.11
yookassa==3.3.0
